+++ src/sputnik/src/js/apis/node/blob/llrt/mod.rs
@@ -1,1 +1,1 @@
-pub mod blob;
+pub mod polyfill;
